Move `ucontext` related functions to their own namespace
